nv/encode: use a reflect.Kind: dataTypes map to verify type
Temporarily have to skip byte encoding tests since byte is an alias to uint8.
This alias is at the spec level as opposed to other types of user/runtime
defined aliases. Because of the special alias type, the runtime cannot
distinguish between uint8 and byte ditto their slice/arrays. Differentiating is
left to the user, which we will implement next.