nv: support byte/uint8 type specification via field tag
Since byte/[]byte are indistinguishable from uint8/[]uint8 we have to provide a
means for users to specify byte/[]byte. We do so in the same way that the xdr
package does, by allowing a `byte` field in a tag. We also allow going from
byte/[]byte to uint8/[]uint8 by specifying `uint8` in the tag since it seems
useful.