fix nvlist array test name
The test name was supposed to encode both the length of the array and the number
of items per array index. The former via the length of the slice returned from
splitting along ';' characters. And the latter by reading the item/number.

We were actually just repeating the length of the array. This worked because
each list had 2 elements, and there were 2 lists. So the name generation has
been fixed, and the test case has been expanded to contain lists of different
sizes.