Add support for Extra fields
If a target struct field is an interface{} or a map and tagged with
"extra (e.g. "nv:foo,extra"), it will be treated as the catch-all map
for nvlist fields not explicitly in the target. Only one struct field
will be used if more than one are tagged "extra".