Shorten acomm.NewRequest func signature and add convenient setters
The function signature was getting out of hand. All of the values could
be set directly on the exported properties. Shortened NewRequest to just
take a task (since that's required) and added a few convenience methods
for setting the ResponseHook and StreamURL from a string, and Args from
an interface. All of those could be set directly as well.

Additionally, regenerated all READMEs