pkg/kv: correctly check etcd Lock contents
A case of `||` when `&&` was actually wanted. If the Lock contents were not
`locked=true` the if-statement would short circuit and return an error.

This passed the tests because a non-existent key was used in the tests for the
positive case, and when testing negative cases an error was expected so yay.
There were not previously any false-negative tests.

The false-negative is tested by testing within pkg/kv/etcd since it knows the
implementation details.