providers/dhcp: make Mock actually work
It was just plainly not working. Also had a nil dereference induced panic.

When a user passes in an empty ip address it gets passed to net.ParseIP, which
returns a nil []byte{}. We  then pass the nil to ipInRange, which will happily
go on to dereference it when doing `ipToU32`.

This situation is now avoided by guarding against the empty string ip
representation.