Improve rollback_clone
SOURCE can now also be just a dataset.
If it is, rollback_clone will try to use the latest snapshot of it.
If no snapshots exist rollback_clone will create a new one named after DESTINATION

As an interesting side effect, if you provide a SOURCE name
that is a full snapshot name (e.g. data/datasets/foo@bar) and
the specified snapshot name doesn't exist but the dataset does,
rollback_clone will create that snapshot and use it.

I've included a script I used to verify the behavior.