providers/kv: lazy init KV.kv
cerana providers/services should be available even if their dependencies are
not. This allows for providers/kv to be up even if the KV store is not yet. This
was only a problem on initial connection, once the connection is established any
communication errors are returned to clients and the provider continues just
fine.