providers/kv: mock: guard against kv down
Especially important for Mock when being run with a low number of GOMAXPROCS
(travis) and running a really quick test like dhcp.TestConfig.

Even though kv.Mock ensures that the kv is up we were still fire'ing off a
goroutine to do the checking. When GOMAXPROCS=1 or maybe 2 the goroutine may not
be scheduled soon enough to populate KV.kv and then when the fast test is done
and the suite is cleaning up by calling Mock.Clean KV.kv is nil and a panic
happens.