cmd/dhcp-provider: call LoadConfig on the embedded provider.Config
cmd/dhcp-provider is no longer getting the configuration via cli so there is
nothing to parse/validate above what provider.Config does. Unfortunately, the
dhcp.Config struct is used by the provider so we can't just get rid of it.

There are some coming changes to dhcp provider that can do the work of getting
rid of dhcp.Config fields.