Call errors.ResetStack on resp.Error
When a response is received by an acomm.Tracker, it is wrapped to
associate the request id. Unfortunately, this creates an identical stack
trace for each, just the response handling goroutine. Using
errors.ResetStack on it once back in the main code gives a more useful
trace to find out where that resp.Error is gotten.