acomm - Log at Debug level instead of Error level
Some errors, such as those from deferred calls or ones generated in
goroutines, can't make it out to the package caller but can't be handled
either. It is up to the caller to verify and handle the visible result.
Debug logging can be turned on to aid in manual investigation.